osutil/disks: move CandidateByLabelPath to labels_linux.go
So we can compile for Mac (os.ReadDir not supported for it).
